Added a "Copy Post Text" action to PostWidget context menu. convert username to lower case before checks
svn path=/trunk/extragear/network/choqok/; revision=1016272

+++ libchoqok/ui/composerwidget.cpp
@@ -75,7 +75,7 @@ void ComposerWidget::submitPost( const QString &txt )
     editor()->setEnabled(false);
     QString text = txt;
     if( currentAccount()->microblog()->postCharLimit() &&
-       text.size() > currentAccount()->microblog()->postCharLimit() )
+       text.size() > (int)currentAccount()->microblog()->postCharLimit() )
         text = Choqok::ShortenManager::self()->parseText(text);
     Choqok::Post *ps = new Choqok::Post;
     ps->content = text;
